+++ src/ext/java/org/opentripplanner/ext/fares/model/FareContainer.java
@@ -1,3 +1,0 @@
-package org.opentripplanner.ext.fares.model;
-
-public record FareContainer(String id, String name) {}