+++ src/ext/java/org/opentripplanner/ext/flex/FlexibleTransitLeg.java
@@ -14,6 +14,7 @@ import org.opentripplanner.framework.lang.DoubleUtils;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
 import org.opentripplanner.model.BookingInfo;
 import org.opentripplanner.model.PickDrop;
+import org.opentripplanner.model.fare.FareProductUse;
 import org.opentripplanner.model.plan.Leg;
 import org.opentripplanner.model.plan.Place;
 import org.opentripplanner.model.plan.StopArrival;
@@ -42,6 +43,7 @@ public class FlexibleTransitLeg implements TransitLeg {
   private final Set<TransitAlert> transitAlerts = new HashSet<>();
 
   private final int generalizedCost;
+  private List<FareProductUse> fareProducts;
 
   public FlexibleTransitLeg(
     FlexTripEdge flexTripEdge,
@@ -201,6 +203,16 @@ public class FlexibleTransitLeg implements TransitLeg {
     }
 
     return copy;
+  }
+
+  @Override
+  public void setFareProducts(List<FareProductUse> products) {
+    this.fareProducts = List.copyOf(products);
+  }
+
+  @Override
+  public List<FareProductUse> fareProducts() {
+    return fareProducts;
   }
 
   /**