+++ src/main/java/org/opentripplanner/routing/algorithm/mapping/StatesToWalkStepsMapper.java
@@ -413,10 +413,11 @@ public class StatesToWalkStepsMapper {
       if (!(out instanceof StreetEdge)) {
         continue;
       }
-      State outState = out.traverse(state.getBackState());
-      if (outState == null) {
+      var outStates = out.traverse(state.getBackState());
+      if (State.isEmpty(outStates)) {
         continue;
       }
+      var outState = outStates[0];
       if (!outState.getBackMode().equals(requestedMode)) {
         //walking a bike, so, not really an exit
         continue;
@@ -429,8 +430,10 @@ public class StatesToWalkStepsMapper {
       Vertex tov = outState.getVertex();
       boolean found = false;
       for (Edge out2 : tov.getOutgoing()) {
-        State outState2 = out2.traverse(outState);
-        if (outState2 != null && !Objects.equals(outState2.getBackMode(), requestedMode)) {
+        var outStates2 = out2.traverse(outState);
+        if (
+          !State.isEmpty(outStates2) && !Objects.equals(outStates2[0].getBackMode(), requestedMode)
+        ) {
           // walking a bike, so, not really an exit
           continue;
         }