+++ src/main/java/org/opentripplanner/street/model/edge/Edge.java
@@ -77,14 +77,6 @@ public abstract class Edge implements AStarEdge<State, Edge, Vertex>, Serializab
   }
 
   /**
-   * Traverse this edge.
-   *
-   * @param s0 The State coming into the edge.
-   * @return The State upon exiting the edge.
-   */
-  public abstract State traverse(State s0);
-
-  /**
    * Returns the default name of the edge
    */
   public String getDefaultName() {