+++ src/main/java/org/opentripplanner/street/model/edge/VehicleParkingEdge.java
@@ -50,9 +50,9 @@ public class VehicleParkingEdge extends Edge {
   }
 
   @Override
-  public State traverse(State s0) {
+  public State[] traverse(State s0) {
     if (!s0.getRequest().mode().includesParking()) {
-      return null;
+      return State.empty();
     }
 
     if (s0.getRequest().arriveBy()) {
@@ -82,9 +82,9 @@ public class VehicleParkingEdge extends Edge {
     return 0;
   }
 
-  protected State traverseUnPark(State s0) {
+  protected State[] traverseUnPark(State s0) {
     if (s0.getNonTransitMode() != TraverseMode.WALK || !s0.isVehicleParked()) {
-      return null;
+      return State.empty();
     }
 
     StreetMode streetMode = s0.getRequest().mode();
@@ -96,11 +96,11 @@ public class VehicleParkingEdge extends Edge {
       final CarPreferences car = s0.getPreferences().car();
       return traverseUnPark(s0, car.parkCost(), car.parkTime(), TraverseMode.CAR);
     } else {
-      return null;
+      return State.empty();
     }
   }
 
-  private State traverseUnPark(State s0, int parkingCost, int parkingTime, TraverseMode mode) {
+  private State[] traverseUnPark(State s0, int parkingCost, int parkingTime, TraverseMode mode) {
     final StreetSearchRequest request = s0.getRequest();
     if (
       !vehicleParking.hasSpacesAvailable(
@@ -109,7 +109,7 @@ public class VehicleParkingEdge extends Edge {
         request.parking().useAvailabilityInformation()
       )
     ) {
-      return null;
+      return State.empty();
     }
 
     StateEditor s0e = s0.edit(this);
@@ -119,32 +119,32 @@ public class VehicleParkingEdge extends Edge {
 
     addUnpreferredTagCost(request.parking(), s0e);
 
-    return s0e.makeState();
+    return s0e.makeStateArray();
   }
 
-  private State traversePark(State s0) {
+  private State[] traversePark(State s0) {
     StreetMode streetMode = s0.getRequest().mode();
     RoutingPreferences preferences = s0.getPreferences();
 
     if (!streetMode.includesWalking() || s0.isVehicleParked()) {
-      return null;
+      return State.empty();
     }
 
     if (streetMode.includesBiking()) {
       // Parking a rented bike is not allowed
       if (s0.isRentingVehicle()) {
-        return null;
+        return State.empty();
       }
 
       return traversePark(s0, preferences.bike().parkCost(), preferences.bike().parkTime());
     } else if (streetMode.includesDriving()) {
       return traversePark(s0, preferences.car().parkCost(), preferences.car().parkTime());
     } else {
-      return null;
+      return State.empty();
     }
   }
 
-  private State traversePark(State s0, int parkingCost, int parkingTime) {
+  private State[] traversePark(State s0, int parkingCost, int parkingTime) {
     if (
       !vehicleParking.hasSpacesAvailable(
         s0.getNonTransitMode(),
@@ -152,7 +152,7 @@ public class VehicleParkingEdge extends Edge {
         s0.getRequest().parking().useAvailabilityInformation()
       )
     ) {
-      return null;
+      return State.empty();
     }
 
     StateEditor s0e = s0.edit(this);
@@ -162,7 +162,7 @@ public class VehicleParkingEdge extends Edge {
 
     addUnpreferredTagCost(s0.getRequest().parking(), s0e);
 
-    return s0e.makeState();
+    return s0e.makeStateArray();
   }
 
   private void addUnpreferredTagCost(VehicleParkingRequest req, StateEditor s0e) {