+++ src/main/java/org/opentripplanner/updater/spi/PollingGraphUpdater.java
@@ -62,7 +62,7 @@ public abstract class PollingGraphUpdater implements GraphUpdater {
         } catch (InterruptedException e) {
           throw e;
         } catch (Exception e) {
-          LOG.error("Error while running polling updater of type {}", configRef, e);
+          LOG.error("Error while running polling updater {}", this, e);
           // TODO Should we cancel the task? Or after n consecutive failures? cancel();
         } finally {
           primed = true;
@@ -70,8 +70,11 @@ public abstract class PollingGraphUpdater implements GraphUpdater {
         Thread.sleep(pollingPeriodSeconds * 1000);
       }
     } catch (InterruptedException e) {
-      // When updater is interrupted
-      LOG.error("Polling updater {} was interrupted and is stopping.", this.getClass().getName());
+      Thread.currentThread().interrupt();
+      LOG.info(
+        "OTP is shutting down, polling updater {} was interrupted and is stopping.",
+        this.getClass().getName()
+      );
     }
   }