+++ docs/Changelog.md
@@ -10,6 +10,14 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Constant speed street routing [#5057](https://github.com/opentripplanner/OpenTripPlanner/pull/5057)
 - Remove dead build configuration parameter (extraEdgesStopPlatformLink) [#5080](https://github.com/opentripplanner/OpenTripPlanner/pull/5080)
 - Configure HTTP WebServer Transaction timeouts [#5047](https://github.com/opentripplanner/OpenTripPlanner/pull/5047)
+- Add Fares v2 to GraphQL API [#4917](https://github.com/opentripplanner/OpenTripPlanner/pull/4917)
+- Improve Graph updaters shutdown [#5092](https://github.com/opentripplanner/OpenTripPlanner/pull/5092)
+- Add deduplicated  to geocoding sandbox API [#5084](https://github.com/opentripplanner/OpenTripPlanner/pull/5084)
+- Fix flex timeshift [#5063](https://github.com/opentripplanner/OpenTripPlanner/pull/5063)
+- Merge norway traversal calculator into default [#5106](https://github.com/opentripplanner/OpenTripPlanner/pull/5106)
+- Use correct GTFS sequence number in vehicle position matcher [#5090](https://github.com/opentripplanner/OpenTripPlanner/pull/5090)
+- Fix SIRI ET PubSub updater shutdown [#5104](https://github.com/opentripplanner/OpenTripPlanner/pull/5104)
+- OSM: Break out of processing a malformed level map [#5096](https://github.com/opentripplanner/OpenTripPlanner/pull/5096)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)